+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2020-03-09
+
 ### Added
 
 - New `Option.collect` constructor to create an `Option[Tuple[T, ...]]`
@@ -81,7 +83,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic benchmarks for analyzing performance
 - Apache license
 
-[Unreleased]: https://github.com/mplanchard/safetywrap/compare/v1.3.1...HEAD
+[Unreleased]: https://github.com/mplanchard/safetywrap/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/mplanchard/safetywrap/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/mplanchard/safetywrap/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/mplanchard/safetywrap/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/mplanchard/safetywrap/compare/v1.1.0...v1.2.0